阶段 D1+D2：GBK 字形查找统一 + 构建卫生告警清零
D1 GBK 解码统一：
- 新增 System/utils/gbk16.{c,h}: gbk16_glyph(b1,b2) 统一范围检查(0xA1-0xF7/0xA1-0xFE)与字库边界(fonts_hzk_end)，修复 UICore.h 下溢与 vGL 上溢缺失
- UICore.h draw_char_GBK16 / kcasporing_gl.c vGL_putChar16 改用共享函数（渲染后端各自保留）

D2 构建卫生：
- System 加 -Wall -Wextra（+ -fstack-protector-strong 栈保护，新增 stack_chk.c 提供 __stack_chk_guard/fail）
- 修复全部可修告警（约 45 处）：main.c format/返回类型、SysIRQ volatile、SystemFs unused、SystemUI 死变量、basic_api 符号比较/悬空 else、port.c %p、warpper、diskio format、CrashLog 缓冲 64->320、UICore.h/cpp 十余处、init.c、threaded.h register、iostream_sub.h format
- 抑制无害类：-Wno-multi-line/-Wno-unused-parameter/-Wno-main(裸机)/-Wno-strict-aliasing(giac 头)/-Wno-unknown-pragmas
- 删除 OSLoader services.c.bak/boot.c.bak；include 目录验证无污染
- 最终全量编译 0 警告

+++ Libs/src/khicas/iostream_sub.h
@@ -59,7 +59,7 @@ inline stdostream & operator << (stdostream & os,double d){
 
 inline stdostream & operator << (stdostream & os,clock_t t){
   char buf[32];
-  sprintf(buf,"%u",t);
+  sprintf(buf,"%lu",(unsigned long)t);
   dConsolePut(buf);
   return os;
 }
